requirements.txt
pyyaml==6.0.1
mccabe==0.7.0
Jinja2==3.1.2
coverage==7.2.7
PyGithub==1.58.1
Flask==2.3.2
Werkzeug==2.3.6